+++ src/ripple/nodestore/backend/LevelDBFactory.cpp
@@ -108,18 +108,27 @@ public:
 
     ~LevelDBBackend()
     {
-        if (m_deletePath)
-        {
-            m_db.reset();
-            boost::filesystem::path dir = m_name;
-            boost::filesystem::remove_all (dir);
-        }
+        close();
     }
 
     std::string
     getName()
     {
         return m_name;
+    }
+
+    void
+    close() override
+    {
+        if (m_db)
+        {
+            m_db.reset();
+            if (m_deletePath)
+            {
+                boost::filesystem::path dir = m_name;
+                boost::filesystem::remove_all (dir);
+            }
+        }
     }
 
     //--------------------------------------------------------------------------
@@ -257,6 +266,11 @@ public:
     writeBatch (Batch const& batch)
     {
         storeBatch (batch);
+    }
+
+    void
+    verify() override
+    {
     }
 };
